+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 ### 2.3.0
 - Add default params to `FreArgException` 
+- Refactor FREArray
 
 ### 2.2.0
 - Add `[DateTime]` support to `FREArray`